backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
pymongo==4.7.2
PyJWT==2.9.0
passlib==1.7.4
bcrypt==4.1.2
pdfminer.six==20231228
gunicorn==21.2.0
